Add general url resolution
